// frontend/src/api/generated/models/system-parameter-category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * パラメータカテゴリ
 * @export
 * @enum {string}
 */

export const SystemParameterCategory = {
    Inventory: 'INVENTORY',
    Outbound: 'OUTBOUND',
    Inbound: 'INBOUND',
    System: 'SYSTEM',
    Security: 'SECURITY'
} as const;

export type SystemParameterCategory = typeof SystemParameterCategory[keyof typeof SystemParameterCategory];